examples/crewai/requirements.txt
crewai
crewai-tools
agentops
python-dotenv
pymarkdown==0.1.4
pymarkdownlnt==0.9.30
pymdown-extensions==10.15